@@ -1,6 +1,6 @@
 package com.appeaser.sublimepickerlibrary.common;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.appeaser.sublimepickerlibrary.datepicker.SublimeDatePicker;
