@@ -18,7 +18,7 @@ package com.appeaser.sublimepickerlibrary.helpers;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import com.appeaser.sublimepickerlibrary.datepicker.SelectedDate;
